+++ src/services/Api.php
@@ -16,6 +16,7 @@ use workingconcept\cloudflare\helpers\UrlHelper;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\RequestException;
+use GuzzleHttp\Exception\GuzzleException;
 use Craft;
 use craft\base\Component;
 use craft\helpers\Json;
@@ -27,7 +28,10 @@ use craft\helpers\Json;
  */
 class Api extends Component
 {
-    const API_BASE_URL = 'https://api.cloudflare.com/client/v4/';
+    /**
+     * Base Cloudflare API URL.
+     */
+    public const API_BASE_URL = 'https://api.cloudflare.com/client/v4/';
 
     /**
      * @var array
@@ -35,7 +39,7 @@ class Api extends Component
     protected $responseItems;
 
     /**
-     * @var \GuzzleHttp\Client
+     * @var Client
      */
     private $_client;
 
@@ -50,10 +54,9 @@ class Api extends Component
      *
      * @return Client|null
      */
-    public function getClient()
+    public function getClient(): ?Client
     {
-        if ($this->_client === null && ConfigHelper::isConfigured())
-        {
+        if ($this->_client === null && ConfigHelper::isConfigured()) {
             $this->_client = Craft::createGuzzleClient([
                 'base_uri' => self::API_BASE_URL,
                 'headers' => $this->_getClientHeaders(),
@@ -69,11 +72,11 @@ class Api extends Component
      * Returns true if provided credentials can successfully make API calls.
      *
      * @return bool
+     * @throws GuzzleException
      */
     public function verifyConnection(): bool
     {
-        if ($this->getClient() === null)
-        {
+        if ($this->getClient() === null) {
             return false;
         }
 
@@ -81,8 +84,7 @@ class Api extends Component
         $testUri = $authType === Settings::AUTH_TYPE_KEY ?
             'zones?per_page=1' : 'user/tokens/verify';
 
-        try
-        {
+        try {
             $response = $this->getClient()->get($testUri);
             $responseContents = Json::decode(
                 $response->getBody()->getContents(),
@@ -93,31 +95,24 @@ class Api extends Component
             $success = $response->getStatusCode() === 200
                 && $responseContents->success;
 
-            if ($success)
-            {
+            if ($success) {
                 return true;
             }
 
-            if (isset($responseContents->errors))
-            {
+            if (isset($responseContents->errors)) {
                 $this->_connectionErrors = $responseContents->errors;
 
                 Craft::info(sprintf(
                     'Connection test failed: %s',
                     Json::encode($responseContents->errors)
                 ), 'cloudflare');
-            }
-            else
-            {
+            } else {
                 Craft::info('Connection test failed.', 'cloudflare');
             }
-        }
-        catch (RequestException $exception)
-        {
+        } catch (RequestException $exception) {
             $reason = $this->_getExceptionReason($exception);
 
-            if (($data = Json::decode($reason, false)) && isset($data->errors))
-            {
+            if (($data = Json::decode($reason, false)) && isset($data->errors)) {
                 $this->_connectionErrors = $data->errors;
             }
 
@@ -131,6 +126,8 @@ class Api extends Component
     }
 
     /**
+     * Returns an array of connection errors.
+     *
      * @return array
      */
     public function getConnectionErrors(): array
@@ -139,17 +136,18 @@ class Api extends Component
     }
 
     /**
+     * Returns `true` if we’re able to list available zones.
+     *
      * @return bool
+     * @throws GuzzleException
      */
     public function canListZones(): bool
     {
-        if ( ! $this->getClient())
-        {
+        if ( ! $this->getClient()) {
             return false;
         }
 
-        if ($response = $this->_getPagedZones(0, 2))
-        {
+        if ($response = $this->_getPagedZones(0, 2)) {
             return count($response->result) > 0;
         }
 
@@ -161,11 +159,11 @@ class Api extends Component
      * https://api.cloudflare.com/#zone-list-zones
      *
      * @return array|null zones from response.result (combined if there was pagination)
+     * @throws GuzzleException
      */
-    public function getZones()
+    public function getZones(): ?array
     {
-        if ( ! $this->getClient())
-        {
+        if ( ! $this->getClient()) {
             return null;
         }
 
@@ -175,24 +173,18 @@ class Api extends Component
         $totalPages  = 1; // temporary
         $perPage     = 50;
 
-        while ($currentPage < $totalPages)
-        {
+        while ($currentPage < $totalPages) {
             $currentPage++;
 
-            if ($response = $this->_getPagedZones($currentPage, $perPage))
-            {
-                if (count($response->result) > 0)
-                {
+            if ($response = $this->_getPagedZones($currentPage, $perPage)) {
+                if (count($response->result) > 0) {
                     $totalRecords = $response->result_info->total_count;
                     $totalPages = ceil($totalRecords / $perPage);
 
-                    foreach ($response->result as $item)
-                    {
+                    foreach ($response->result as $item) {
                         $this->responseItems[] = $item;
                     }
-                }
-                else
-                {
+                } else {
                     return [];
                 }
             }
@@ -207,22 +199,21 @@ class Api extends Component
      *
      * @param string $zoneId
      * @return object|null
+     * @throws GuzzleException
      */
-    public function getZoneById($zoneId)
+    public function getZoneById(string $zoneId): ?object
     {
-        if (! $this->getClient())
-        {
+        if (! $this->getClient()) {
             return null;
         }
 
-        try
-        {
+        try {
             $response = $this->getClient()->get(sprintf(
                 'zones/%s',
                 $zoneId
             ));
 
-            if ( ! $response->getStatusCode() == 200)
+            if ( ! $response->getStatusCode() === 200)
             {
                 Craft::info(sprintf(
                     'Request failed: %s',
@@ -231,9 +222,7 @@ class Api extends Component
 
                 return null;
             }
-        }
-        catch(RequestException $exception)
-        {
+        } catch(RequestException $exception) {
             Craft::info(sprintf(
                     'Zone request failed: %s',
                     $this->_getExceptionReason($exception)
@@ -244,10 +233,8 @@ class Api extends Component
             return null;
         }
 
-        Craft::info(sprintf(
-                'Retrieved zone #%s',
-                $zoneId
-            ),
+        Craft::info(
+            sprintf('Retrieved zone #%s', $zoneId),
             'cloudflare'
         );
 
@@ -259,17 +246,16 @@ class Api extends Component
      * Purge the entire zone cache.
      * https://api.cloudflare.com/#zone-purge-all-files
      *
-     * @return object|null Cloudflare's response
+     * @return object|null Cloudflare’s response
+     * @throws GuzzleException
      */
     public function purgeZoneCache()
     {
-        if ( ! $this->getClient())
-        {
+        if ( ! $this->getClient()) {
             return null;
         }
 
-        try
-        {
+        try {
             $response = $this->getClient()->delete(sprintf(
                     'zones/%s/purge_cache',
                     ConfigHelper::getParsedSetting('zone')
@@ -279,8 +265,7 @@ class Api extends Component
 
             $responseBody = Json::decode($response->getBody(), false);
 
-            if ( ! $response->getStatusCode() == 200)
-            {
+            if ( ! $response->getStatusCode() === 200) {
                 Craft::info(sprintf(
                     'Zone purge request failed: %s',
                     Json::encode($responseBody)
@@ -293,50 +278,42 @@ class Api extends Component
                 ];
             }
 
-            Craft::info(sprintf(
-                'Purged entire zone cache (%s)',
-                $responseBody->result->id
-            ), 'cloudflare');
+            Craft::info(
+                sprintf('Purged entire zone cache (%s)', $responseBody->result->id),
+                'cloudflare'
+            );
 
             return $responseBody;
-        }
-        catch(ClientException $exception)
-        {
-            return $this->_handleApiException($exception, 'zone purge');
-        }
-        catch(RequestException $exception)
-        {
+        } catch(ClientException | RequestException $exception) {
             return $this->_handleApiException($exception, 'zone purge');
         }
     }
 
     /**
-     * Clear specific URLs in Cloudflare's cache.
+     * Clear specific URLs in Cloudflare’s cache.
      * https://api.cloudflare.com/#zone-purge-individual-files-by-url-and-cache-tags
      *
-     * @param  array  $urls  array of absolute URLs
+     * @param array $urls array of absolute URLs
      *
      * @return mixed|null  API response data or null
+     * @throws GuzzleException
      */
     public function purgeUrls(array $urls = [])
     {
-        if ( ! $this->getClient())
-        {
+        if ( ! $this->getClient()) {
             return null;
         }
 
         $urls = UrlHelper::prepUrls($urls);
 
-        // don't do anything if URLs are missing
-        if (count($urls) === 0)
-        {
+        // don’t do anything if URLs are missing
+        if (count($urls) === 0) {
             return $this->_failureResponse(
                 'Cannot purge; no valid URLs.'
             );
         }
 
-        try
-        {
+        try {
             $response = $this->getClient()->delete(sprintf(
                     'zones/%s/purge_cache',
                     ConfigHelper::getParsedSetting('zone')
@@ -346,8 +323,7 @@ class Api extends Component
 
             $responseBody = Json::decode($response->getBody(), false);
 
-            if ( ! $response->getStatusCode() == 200)
-            {
+            if ( ! $response->getStatusCode() === 200) {
                 Craft::info(sprintf(
                     'Request failed: %s',
                     Json::encode($responseBody)
@@ -370,18 +346,13 @@ class Api extends Component
 
             return $responseBody;
         }
-        catch(ClientException $exception)
-        {
-            return $this->_handleApiException($exception, 'URL purge', $urls);
-        }
-        catch(RequestException $exception)
-        {
+        catch(ClientException | RequestException $exception) {
             return $this->_handleApiException($exception, 'URL purge', $urls);
         }
     }
 
     /**
-     * Get Cloudflare's base API URL.
+     * Get Cloudflare’s base API URL.
      *
      * @return string
      */
@@ -399,19 +370,16 @@ class Api extends Component
      *
      * @return \stdClass with populated `result` property array
      */
-    private function _handleApiException($exception, $action, $urls = []): \stdClass
+    private function _handleApiException($exception, string $action, $urls = []): \stdClass
     {
-        if ($responseBody = Json::decode($exception->getResponse()->getBody(), false))
-        {
+        if ($responseBody = Json::decode($exception->getResponse()->getBody(), false)) {
             $message = "${action} failed.\n";
 
-            if ($urls)
-            {
+            if ($urls) {
                 $message .= '- urls: ' . implode(',', $urls) . "\n";
             }
 
-            foreach ($responseBody->errors as $error)
-            {
+            foreach ($responseBody->errors as $error) {
                 $message .= "- error code {$error->code}: " . $error->message . "\n";
             }
 
@@ -438,11 +406,11 @@ class Api extends Component
      * @param int $perPage
      *
      * @return \stdClass|null
+     * @throws GuzzleException
      */
     private function _getPagedZones($page = 1, $perPage = 50)
     {
-        if ( ! $this->getClient())
-        {
+        if ( ! $this->getClient()) {
             return null;
         }
 
@@ -451,15 +419,13 @@ class Api extends Component
             $page
         ), 'cloudflare');
 
-        try
-        {
+        try {
             $response = $this->getClient()->get(sprintf(
                 'zones?per_page=%d',
                 $perPage
             ));
 
-            if ( ! $response->getStatusCode() == 200)
-            {
+            if ( ! $response->getStatusCode() === 200) {
                 return $this->_failureResponse(sprintf(
                     'Request failed: %s',
                     $response->getBody()
@@ -469,9 +435,7 @@ class Api extends Component
             Craft::info('Retrieved zones.', 'cloudflare');
 
             return Json::decode($response->getBody(), false);
-        }
-        catch (RequestException $exception)
-        {
+        } catch (RequestException $exception) {
             return $this->_failureResponse(sprintf(
                 'Request failed: %s',
                 $this->_getExceptionReason($exception)
@@ -490,16 +454,13 @@ class Api extends Component
 
         $headers = [
             'Content-Type' => 'application/json',
-            'Accept'       => 'application/json'
+            'Accept' => 'application/json'
         ];
 
-        if ($authType === Settings::AUTH_TYPE_KEY)
-        {
+        if ($authType === Settings::AUTH_TYPE_KEY) {
             $headers['X-Auth-Key'] = ConfigHelper::getParsedSetting('apiKey');
             $headers['X-Auth-Email'] = ConfigHelper::getParsedSetting('email');
-        }
-        elseif ($authType === Settings::AUTH_TYPE_TOKEN)
-        {
+        } elseif ($authType === Settings::AUTH_TYPE_TOKEN) {
             $headers['Authorization'] = sprintf(
                 'Bearer %s',
                 ConfigHelper::getParsedSetting('apiToken')
@@ -516,7 +477,7 @@ class Api extends Component
      *
      * @return object
      */
-    private function _failureResponse($message)
+    private function _failureResponse($message): object
     {
         Craft::error($message, 'cloudflare');
 
@@ -530,7 +491,7 @@ class Api extends Component
     /**
      * Returns a string for the request exception that can be used for logging.
      *
-     * @param  \GuzzleHttp\Exception\RequestException  $exception
+     * @param RequestException  $exception
      *
      * @return string
      */