+++ src/variables/CloudflareVariable.php
@@ -11,6 +11,7 @@
 namespace workingconcept\cloudflare\variables;
 
 use workingconcept\cloudflare\Cloudflare;
+use craft\base\Model;
 
 /**
  * @author    Working Concept
@@ -24,9 +25,9 @@ class CloudflareVariable
      *
      * @return array|null
      */
-    public function getZones()
+    public function getZones(): ?array
     {
-        return Cloudflare::$plugin->api->getZones();
+        return Cloudflare::getInstance()->api->getZones();
     }
 
     /**
@@ -38,10 +39,8 @@ class CloudflareVariable
     {
         $options = [];
 
-        if ($zones = $this->getZones())
-        {
-            foreach ($zones as $zone)
-            {
+        if ($zones = $this->getZones()) {
+            foreach ($zones as $zone) {
                 $options[$zone->id] = $zone->name;
             }
         }
@@ -56,14 +55,14 @@ class CloudflareVariable
      */
     public function getRulesForTable(): array
     {
-        return Cloudflare::$plugin->rules->getRulesForTable();
+        return Cloudflare::getInstance()->rules->getRulesForTable();
     }
 
     /**
-     * @return bool|\craft\base\Model|null
+     * @return bool|Model|null
      */
     public function getSettings()
     {
-        return Cloudflare::$plugin->getSettings();
+        return Cloudflare::getInstance()->getSettings();
     }
 }